+++ apps/docs/content/components/index.ts
@@ -16,3 +16,4 @@ export * from "./link";
 export * from "./progress";
 export * from "./user";
 export * from "./skeleton";
+export * from "./snippet";