+++ apps/docs/content/components/snippet/custom-tooltip.ts
@@ -1,0 +1,25 @@
+const App = `import {Snippet} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Snippet
+      tooltipProps={{
+        color: "foreground",
+        content: "Copy this snippet",
+        disableAnimation: true,
+        placement: "right",
+        closeDelay: 0
+      }}
+    >
+      npm install @nextui-org/react
+    </Snippet>
+  );
+}`;
+
+const react = {
+  "/App.jsx": App,
+};
+
+export default {
+  ...react,
+};