+++ apps/docs/content/components/snippet/index.ts
@@ -1,0 +1,21 @@
+import usage from "./usage";
+import sizes from "./sizes";
+import colors from "./colors";
+import variants from "./variants";
+import symbol from "./symbol";
+import withoutCopy from "./without-copy";
+import customTooltip from "./custom-tooltip";
+import multiline from "./multiline";
+import customIcons from "./custom-icons";
+
+export const snippetContent = {
+  usage,
+  sizes,
+  colors,
+  variants,
+  symbol,
+  withoutCopy,
+  customTooltip,
+  multiline,
+  customIcons,
+};