+++ apps/docs/content/components/snippet/without-copy.ts
@@ -1,0 +1,17 @@
+const App = `import {Snippet} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Snippet hideCopyButton color="primary" variant="solid">
+      npm install @nextui-org/react
+    </Snippet>
+  );
+}`;
+
+const react = {
+  "/App.jsx": App,
+};
+
+export default {
+  ...react,
+};