+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @nextui-org/input
 
+## 0.0.0-dev-v2-20230528191554
+
+### Patch Changes
+
+- Snippet multiline copy fixed and styles improved
+- Updated dependencies
+  - @nextui-org/shared-icons@0.0.0-dev-v2-20230528191554
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230528191554
+  - @nextui-org/use-aria-field@0.0.0-dev-v2-20230528191554
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230528191554
+  - @nextui-org/system@0.0.0-dev-v2-20230528191554
+  - @nextui-org/theme@0.0.0-dev-v2-20230528191554
+
 ## 0.0.0-dev-v2-20230528153341
 
 ### Patch Changes